split warrior.js into multiple files to make it easier to work with
